backend/requirements.txt
flask
flask-cors
python-dotenv
azure-cognitiveservices-vision-computervision
msrest
gunicorn
